refactor: frontend home structure
